fix: messy fix

+++ internal/database/db.go
@@ -24,7 +24,7 @@ func Initialize() {
 func createTables() {
 	userTable := `
     CREATE TABLE IF NOT EXISTS users (
-        id FLOAT PRIMARY KEY AUTOINCREMENT,
+        id INTEGER PRIMARY KEY AUTOINCREMENT,
         name TEXT NOT NULL,
         email TEXT UNIQUE NOT NULL,
         password TEXT NOT NULL
